requirements.txt
pandas
python-dotenv
elasticsearch
sentence-transformers
datasets
pyarrow